Add : Support VP30/31 support. remove some C4018 warning;
git-svn-id: https://mpc-hc.svn.sourceforge.net/svnroot/mpc-hc/trunk@4066 10f7b99b-c216-0410-bff0-8a66a9350fd8

+++ src/filters/muxer/BaseMuxer/BaseMuxerOutputPin.cpp
@@ -278,7 +278,7 @@ void CBaseMuxerRawOutputPin::MuxPacket(const CMediaType& mt, const MuxerPacket* 
 				p += 4;
 				i -= 4;
 
-				if (len > i || len == 1) {
+				if (len > (DWORD)i || len == 1) {
 					len = i;
 					ASSERT(0);
 				}
